+++ bin/src/compilePhase.d.ts
@@ -1,0 +1,4 @@
+import { cceParams } from '.';
+export declare const mainMatch: RegExp;
+export declare const modifiedFile = "__cce_mod__.c";
+export default function compilePhase(params: cceParams, callback: () => any): void;
