+++ bin/src/executePhase.d.ts
@@ -1,0 +1,2 @@
+import { cceParams } from '.';
+export default function executePhase(params: cceParams): void;
