+++ bin/src/index.d.ts
@@ -1,4 +1,5 @@
 #!/usr/bin/env node
+import { ParsedObject } from './parseArgs';
 export declare type Options = {
     compiler?: string;
     compilerArguments?: string;
@@ -9,4 +10,16 @@ export declare type Options = {
     save?: true;
     asIs?: true;
     onlyExecPrints?: true;
+};
+export declare type cceParams = {
+    file: string;
+    compiler: string;
+    compilerArguments: string;
+    executeArguments: ParsedObject;
+    stdin: string;
+    stdout: string;
+    outfile: string;
+    save: boolean;
+    asIs: boolean;
+    onlyExecPrints: boolean;
 };
