+++ bin/src/parseArgs.d.ts
@@ -1,7 +1,6 @@
-declare type ParsedObject = {
+export declare type ParsedObject = {
     parsed: string[];
     error: boolean;
     specials: string[];
 };
 export default function parseArgs(args: string): ParsedObject;
-export {};
