+++ bin/src/validatePhase.d.ts
@@ -1,0 +1,2 @@
+import { cceParams } from '.';
+export default function validatePhase(params: cceParams, callback: () => any): void;
