+++ bin/src/validatePhase.js
@@ -1,0 +1,107 @@
+"use strict";
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
+var __generator = (this && this.__generator) || function (thisArg, body) {
+    var _ = { label: 0, sent: function() { if (t[0] & 1) throw t[1]; return t[1]; }, trys: [], ops: [] }, f, y, t, g;
+    return g = { next: verb(0), "throw": verb(1), "return": verb(2) }, typeof Symbol === "function" && (g[Symbol.iterator] = function() { return this; }), g;
+    function verb(n) { return function (v) { return step([n, v]); }; }
+    function step(op) {
+        if (f) throw new TypeError("Generator is already executing.");
+        while (_) try {
+            if (f = 1, y && (t = op[0] & 2 ? y["return"] : op[0] ? y["throw"] || ((t = y["return"]) && t.call(y), 0) : y.next) && !(t = t.call(y, op[1])).done) return t;
+            if (y = 0, t) op = [op[0] & 2, t.value];
+            switch (op[0]) {
+                case 0: case 1: t = op; break;
+                case 4: _.label++; return { value: op[1], done: false };
+                case 5: _.label++; y = op[1]; op = [0]; continue;
+                case 7: op = _.ops.pop(); _.trys.pop(); continue;
+                default:
+                    if (!(t = _.trys, t = t.length > 0 && t[t.length - 1]) && (op[0] === 6 || op[0] === 2)) { _ = 0; continue; }
+                    if (op[0] === 3 && (!t || (op[1] > t[0] && op[1] < t[3]))) { _.label = op[1]; break; }
+                    if (op[0] === 6 && _.label < t[1]) { _.label = t[1]; t = op; break; }
+                    if (t && _.label < t[2]) { _.label = t[2]; _.ops.push(op); break; }
+                    if (t[2]) _.ops.pop();
+                    _.trys.pop(); continue;
+            }
+            op = body.call(thisArg, _);
+        } catch (e) { op = [6, e]; y = 0; } finally { f = t = 0; }
+        if (op[0] & 5) throw op[1]; return { value: op[0] ? op[1] : void 0, done: true };
+    }
+};
+Object.defineProperty(exports, "__esModule", { value: true });
+var chalk = require("chalk");
+var child_process_1 = require("child_process");
+var fs_1 = require("fs");
+var path = require("path");
+var loading_1 = require("./loading");
+function validatePhase(params, callback) {
+    var _this = this;
+    var file = params.file, onlyExecPrints = params.onlyExecPrints, executeArguments = params.executeArguments, compilerArguments = params.compilerArguments, compiler = params.compiler;
+    var filePath = path.join(process.cwd(), file);
+    var validatingLoader = !onlyExecPrints && (0, loading_1.default)('Validating');
+    if (executeArguments.error) {
+        validatingLoader.error();
+        console.error('error: missing dequote from execute arguments');
+        process.exit(1);
+    }
+    var warnings = [];
+    if (compilerArguments.match(/(-o|--output)/))
+        warnings.push('--compiler-arguments contains an --output option, this could prevent CCE from executing the compiled file, please use the CCE --outfile (-o) option instead');
+    executeArguments.specials.forEach(function (s) {
+        switch (s[0]) {
+            case '<': {
+                warnings.push('--execute-arguments contains redirect <, this will not have any effect, please use CCE --stdin option instead');
+                break;
+            }
+            case '>': {
+                warnings.push('--execute-arguments contains redirect >, this will not have any effect, please use CCE --stdout option instead');
+                break;
+            }
+        }
+    });
+    if (!(0, fs_1.existsSync)(filePath)) {
+        if (!onlyExecPrints)
+            validatingLoader.error();
+        console.error("error: file ".concat(filePath, " doesn't exist"));
+        process.exit(1);
+    }
+    (0, fs_1.stat)(filePath, function (error, stats) {
+        if (error) {
+            if (!onlyExecPrints)
+                validatingLoader.error();
+            console.error("error: there was an issue retrieving information about the file ".concat(filePath));
+            process.exit(1);
+        }
+        if (!stats.isFile()) {
+            if (!onlyExecPrints)
+                validatingLoader.error();
+            console.error("error: ".concat(filePath, " is not a file"));
+            process.exit(1);
+        }
+        (0, child_process_1.exec)("command -v ".concat(compiler), function (error) { return __awaiter(_this, void 0, void 0, function () {
+            return __generator(this, function (_a) {
+                if (error) {
+                    if (!onlyExecPrints)
+                        validatingLoader.error();
+                    console.error("error: compiler ".concat(compiler, " doesn't exist"));
+                    process.exit(1);
+                }
+                if (!onlyExecPrints) {
+                    validatingLoader.done();
+                    warnings.forEach(function (w) { return console.log("".concat(chalk.hex('#a2e')('Warning:'), " ").concat(w)); });
+                }
+                callback();
+                return [2];
+            });
+        }); });
+    });
+}
+exports.default = validatePhase;
+//# sourceMappingURL=validatePhase.js.map
